Implement CRUD operations for photos and pets
The commit introduces Create, Read, Update, and Delete (CRUD) operations on photos and pets in the application. This has been achieved by adding various forms for adding, editing and deleting pets and photos. The commit also enables comments addition on photos, and implements the search functionality in the home page. Authentication checks have not been considered and will be included in future commits. Forms have been used in templates to represent data, and form data is now being saved in the database.